+++ include/zephyr/drivers/uart.h
@@ -468,18 +468,6 @@ __subsystem struct uart_driver_api {
  */
 __syscall int uart_err_check(const struct device *dev);
 
-static inline int z_impl_uart_err_check(const struct device *dev)
-{
-	const struct uart_driver_api *api =
-		(const struct uart_driver_api *)dev->api;
-
-	if (api->err_check == NULL) {
-		return -ENOSYS;
-	}
-
-	return api->err_check(dev);
-}
-
 /**
  * @defgroup uart_polling Polling UART API
  * @{
@@ -505,19 +493,6 @@ static inline int z_impl_uart_err_check(const struct device *dev)
  */
 __syscall int uart_poll_in(const struct device *dev, unsigned char *p_char);
 
-static inline int z_impl_uart_poll_in(const struct device *dev,
-				      unsigned char *p_char)
-{
-	const struct uart_driver_api *api =
-		(const struct uart_driver_api *)dev->api;
-
-	if (api->poll_in == NULL) {
-		return -ENOSYS;
-	}
-
-	return api->poll_in(dev, p_char);
-}
-
 /**
  * @brief Read a 16-bit datum from the device for input.
  *
@@ -539,25 +514,6 @@ static inline int z_impl_uart_poll_in(const struct device *dev,
  */
 __syscall int uart_poll_in_u16(const struct device *dev, uint16_t *p_u16);
 
-static inline int z_impl_uart_poll_in_u16(const struct device *dev,
-					  uint16_t *p_u16)
-{
-#ifdef CONFIG_UART_WIDE_DATA
-	const struct uart_driver_api *api =
-		(const struct uart_driver_api *)dev->api;
-
-	if (api->poll_in_u16 == NULL) {
-		return -ENOSYS;
-	}
-
-	return api->poll_in_u16(dev, p_u16);
-#else
-	ARG_UNUSED(dev);
-	ARG_UNUSED(p_u16);
-	return -ENOTSUP;
-#endif
-}
-
 /**
  * @brief Write a character to the device for output.
  *
@@ -575,15 +531,6 @@ static inline int z_impl_uart_poll_in_u16(const struct device *dev,
 __syscall void uart_poll_out(const struct device *dev,
 			     unsigned char out_char);
 
-static inline void z_impl_uart_poll_out(const struct device *dev,
-					unsigned char out_char)
-{
-	const struct uart_driver_api *api =
-		(const struct uart_driver_api *)dev->api;
-
-	api->poll_out(dev, out_char);
-}
-
 /**
  * @brief Write a 16-bit datum to the device for output.
  *
@@ -599,20 +546,6 @@ static inline void z_impl_uart_poll_out(const struct device *dev,
  * @param out_u16 Wide data to send.
  */
 __syscall void uart_poll_out_u16(const struct device *dev, uint16_t out_u16);
-
-static inline void z_impl_uart_poll_out_u16(const struct device *dev,
-					    uint16_t out_u16)
-{
-#ifdef CONFIG_UART_WIDE_DATA
-	const struct uart_driver_api *api =
-		(const struct uart_driver_api *)dev->api;
-
-	api->poll_out_u16(dev, out_u16);
-#else
-	ARG_UNUSED(dev);
-	ARG_UNUSED(out_u16);
-#endif
-}
 
 /**
  * @}
@@ -635,24 +568,6 @@ static inline void z_impl_uart_poll_out_u16(const struct device *dev,
 __syscall int uart_configure(const struct device *dev,
 			     const struct uart_config *cfg);
 
-static inline int z_impl_uart_configure(const struct device *dev,
-					const struct uart_config *cfg)
-{
-#ifdef CONFIG_UART_USE_RUNTIME_CONFIGURE
-	const struct uart_driver_api *api =
-				(const struct uart_driver_api *)dev->api;
-
-	if (api->configure == NULL) {
-		return -ENOSYS;
-	}
-	return api->configure(dev, cfg);
-#else
-	ARG_UNUSED(dev);
-	ARG_UNUSED(cfg);
-	return -ENOTSUP;
-#endif
-}
-
 /**
  * @brief Get UART configuration.
  *
@@ -669,25 +584,6 @@ static inline int z_impl_uart_configure(const struct device *dev,
  */
 __syscall int uart_config_get(const struct device *dev,
 			      struct uart_config *cfg);
-
-static inline int z_impl_uart_config_get(const struct device *dev,
-					 struct uart_config *cfg)
-{
-#ifdef CONFIG_UART_USE_RUNTIME_CONFIGURE
-	const struct uart_driver_api *api =
-				(const struct uart_driver_api *)dev->api;
-
-	if (api->config_get == NULL) {
-		return -ENOSYS;
-	}
-
-	return api->config_get(dev, cfg);
-#else
-	ARG_UNUSED(dev);
-	ARG_UNUSED(cfg);
-	return -ENOTSUP;
-#endif
-}
 
 /**
  * @addtogroup uart_interrupt
@@ -866,40 +762,12 @@ static inline int uart_fifo_read_u16(const struct device *dev,
  */
 __syscall void uart_irq_tx_enable(const struct device *dev);
 
-static inline void z_impl_uart_irq_tx_enable(const struct device *dev)
-{
-#ifdef CONFIG_UART_INTERRUPT_DRIVEN
-	const struct uart_driver_api *api =
-		(const struct uart_driver_api *)dev->api;
-
-	if (api->irq_tx_enable != NULL) {
-		api->irq_tx_enable(dev);
-	}
-#else
-	ARG_UNUSED(dev);
-#endif
-}
-
 /**
  * @brief Disable TX interrupt in IER.
  *
  * @param dev UART device instance.
  */
 __syscall void uart_irq_tx_disable(const struct device *dev);
-
-static inline void z_impl_uart_irq_tx_disable(const struct device *dev)
-{
-#ifdef CONFIG_UART_INTERRUPT_DRIVEN
-	const struct uart_driver_api *api =
-		(const struct uart_driver_api *)dev->api;
-
-	if (api->irq_tx_disable != NULL) {
-		api->irq_tx_disable(dev);
-	}
-#else
-	ARG_UNUSED(dev);
-#endif
-}
 
 /**
  * @brief Check if UART TX buffer can accept a new char
@@ -943,40 +811,12 @@ static inline int uart_irq_tx_ready(const struct device *dev)
  */
 __syscall void uart_irq_rx_enable(const struct device *dev);
 
-static inline void z_impl_uart_irq_rx_enable(const struct device *dev)
-{
-#ifdef CONFIG_UART_INTERRUPT_DRIVEN
-	const struct uart_driver_api *api =
-		(const struct uart_driver_api *)dev->api;
-
-	if (api->irq_rx_enable != NULL) {
-		api->irq_rx_enable(dev);
-	}
-#else
-	ARG_UNUSED(dev);
-#endif
-}
-
 /**
  * @brief Disable RX interrupt.
  *
  * @param dev UART device instance.
  */
 __syscall void uart_irq_rx_disable(const struct device *dev);
-
-static inline void z_impl_uart_irq_rx_disable(const struct device *dev)
-{
-#ifdef CONFIG_UART_INTERRUPT_DRIVEN
-	const struct uart_driver_api *api =
-		(const struct uart_driver_api *)dev->api;
-
-	if (api->irq_rx_disable != NULL) {
-		api->irq_rx_disable(dev);
-	}
-#else
-	ARG_UNUSED(dev);
-#endif
-}
 
 /**
  * @brief Check if UART TX block finished transmission
@@ -1055,40 +895,12 @@ static inline int uart_irq_rx_ready(const struct device *dev)
  */
 __syscall void uart_irq_err_enable(const struct device *dev);
 
-static inline void z_impl_uart_irq_err_enable(const struct device *dev)
-{
-#ifdef CONFIG_UART_INTERRUPT_DRIVEN
-	const struct uart_driver_api *api =
-		(const struct uart_driver_api *)dev->api;
-
-	if (api->irq_err_enable) {
-		api->irq_err_enable(dev);
-	}
-#else
-	ARG_UNUSED(dev);
-#endif
-}
-
 /**
  * @brief Disable error interrupt.
  *
  * @param dev UART device instance.
  */
 __syscall void uart_irq_err_disable(const struct device *dev);
-
-static inline void z_impl_uart_irq_err_disable(const struct device *dev)
-{
-#ifdef CONFIG_UART_INTERRUPT_DRIVEN
-	const struct uart_driver_api *api =
-		(const struct uart_driver_api *)dev->api;
-
-	if (api->irq_err_disable) {
-		api->irq_err_disable(dev);
-	}
-#else
-	ARG_UNUSED(dev);
-#endif
-}
 
 /**
  * @brief Check if any IRQs is pending.
@@ -1101,22 +913,6 @@ static inline void z_impl_uart_irq_err_disable(const struct device *dev)
  * @retval -ENOTSUP If API is not enabled.
  */
 __syscall int uart_irq_is_pending(const struct device *dev);
-
-static inline int z_impl_uart_irq_is_pending(const struct device *dev)
-{
-#ifdef CONFIG_UART_INTERRUPT_DRIVEN
-	const struct uart_driver_api *api =
-		(const struct uart_driver_api *)dev->api;
-
-	if (api->irq_is_pending == NULL) {
-		return -ENOSYS;
-	}
-	return api->irq_is_pending(dev);
-#else
-	ARG_UNUSED(dev);
-	return -ENOTSUP;
-#endif
-}
 
 /**
  * @brief Start processing interrupts in ISR.
@@ -1144,22 +940,6 @@ static inline int z_impl_uart_irq_is_pending(const struct device *dev)
  * @retval -ENOTSUP If API is not enabled.
  */
 __syscall int uart_irq_update(const struct device *dev);
-
-static inline int z_impl_uart_irq_update(const struct device *dev)
-{
-#ifdef CONFIG_UART_INTERRUPT_DRIVEN
-	const struct uart_driver_api *api =
-		(const struct uart_driver_api *)dev->api;
-
-	if (api->irq_update == NULL) {
-		return -ENOSYS;
-	}
-	return api->irq_update(dev);
-#else
-	ARG_UNUSED(dev);
-	return -ENOTSUP;
-#endif
-}
 
 /**
  * @brief Set the IRQ callback function pointer.
@@ -1283,24 +1063,6 @@ __syscall int uart_tx(const struct device *dev, const uint8_t *buf,
 		      size_t len,
 		      int32_t timeout);
 
-static inline int z_impl_uart_tx(const struct device *dev, const uint8_t *buf,
-				 size_t len, int32_t timeout)
-
-{
-#ifdef CONFIG_UART_ASYNC_API
-	const struct uart_driver_api *api =
-			(const struct uart_driver_api *)dev->api;
-
-	return api->tx(dev, buf, len, timeout);
-#else
-	ARG_UNUSED(dev);
-	ARG_UNUSED(buf);
-	ARG_UNUSED(len);
-	ARG_UNUSED(timeout);
-	return -ENOTSUP;
-#endif
-}
-
 /**
  * @brief Send given number of datum from buffer through UART.
  *
@@ -1321,25 +1083,6 @@ static inline int z_impl_uart_tx(const struct device *dev, const uint8_t *buf,
 __syscall int uart_tx_u16(const struct device *dev, const uint16_t *buf,
 			  size_t len, int32_t timeout);
 
-static inline int z_impl_uart_tx_u16(const struct device *dev,
-				     const uint16_t *buf,
-				     size_t len, int32_t timeout)
-
-{
-#if defined(CONFIG_UART_ASYNC_API) && defined(CONFIG_UART_WIDE_DATA)
-	const struct uart_driver_api *api =
-			(const struct uart_driver_api *)dev->api;
-
-	return api->tx_u16(dev, buf, len, timeout);
-#else
-	ARG_UNUSED(dev);
-	ARG_UNUSED(buf);
-	ARG_UNUSED(len);
-	ARG_UNUSED(timeout);
-	return -ENOTSUP;
-#endif
-}
-
 /**
  * @brief Abort current TX transmission.
  *
@@ -1353,19 +1096,6 @@ static inline int z_impl_uart_tx_u16(const struct device *dev,
  * @retval -errno Other negative errno value in case of failure.
  */
 __syscall int uart_tx_abort(const struct device *dev);
-
-static inline int z_impl_uart_tx_abort(const struct device *dev)
-{
-#ifdef CONFIG_UART_ASYNC_API
-	const struct uart_driver_api *api =
-			(const struct uart_driver_api *)dev->api;
-
-	return api->tx_abort(dev);
-#else
-	ARG_UNUSED(dev);
-	return -ENOTSUP;
-#endif
-}
 
 /**
  * @brief Start receiving data through UART.
@@ -1392,24 +1122,6 @@ __syscall int uart_rx_enable(const struct device *dev, uint8_t *buf,
 			     size_t len,
 			     int32_t timeout);
 
-static inline int z_impl_uart_rx_enable(const struct device *dev,
-					uint8_t *buf,
-					size_t len, int32_t timeout)
-{
-#ifdef CONFIG_UART_ASYNC_API
-	const struct uart_driver_api *api =
-				(const struct uart_driver_api *)dev->api;
-
-	return api->rx_enable(dev, buf, len, timeout);
-#else
-	ARG_UNUSED(dev);
-	ARG_UNUSED(buf);
-	ARG_UNUSED(len);
-	ARG_UNUSED(timeout);
-	return -ENOTSUP;
-#endif
-}
-
 /**
  * @brief Start receiving wide data through UART.
  *
@@ -1433,24 +1145,6 @@ static inline int z_impl_uart_rx_enable(const struct device *dev,
  */
 __syscall int uart_rx_enable_u16(const struct device *dev, uint16_t *buf,
 				 size_t len, int32_t timeout);
-
-static inline int z_impl_uart_rx_enable_u16(const struct device *dev,
-					    uint16_t *buf, size_t len,
-					    int32_t timeout)
-{
-#if defined(CONFIG_UART_ASYNC_API) && defined(CONFIG_UART_WIDE_DATA)
-	const struct uart_driver_api *api =
-				(const struct uart_driver_api *)dev->api;
-
-	return api->rx_enable_u16(dev, buf, len, timeout);
-#else
-	ARG_UNUSED(dev);
-	ARG_UNUSED(buf);
-	ARG_UNUSED(len);
-	ARG_UNUSED(timeout);
-	return -ENOTSUP;
-#endif
-}
 
 /**
  * @brief Provide receive buffer in response to #UART_RX_BUF_REQUEST event.
@@ -1542,19 +1236,6 @@ static inline int uart_rx_buf_rsp_u16(const struct device *dev, uint16_t *buf,
  */
 __syscall int uart_rx_disable(const struct device *dev);
 
-static inline int z_impl_uart_rx_disable(const struct device *dev)
-{
-#ifdef CONFIG_UART_ASYNC_API
-	const struct uart_driver_api *api =
-			(const struct uart_driver_api *)dev->api;
-
-	return api->rx_disable(dev);
-#else
-	ARG_UNUSED(dev);
-	return -ENOTSUP;
-#endif
-}
-
 /**
  * @}
  */
@@ -1574,25 +1255,6 @@ static inline int z_impl_uart_rx_disable(const struct device *dev)
 __syscall int uart_line_ctrl_set(const struct device *dev,
 				 uint32_t ctrl, uint32_t val);
 
-static inline int z_impl_uart_line_ctrl_set(const struct device *dev,
-					    uint32_t ctrl, uint32_t val)
-{
-#ifdef CONFIG_UART_LINE_CTRL
-	const struct uart_driver_api *api =
-		(const struct uart_driver_api *)dev->api;
-
-	if (api->line_ctrl_set == NULL) {
-		return -ENOSYS;
-	}
-	return api->line_ctrl_set(dev, ctrl, val);
-#else
-	ARG_UNUSED(dev);
-	ARG_UNUSED(ctrl);
-	ARG_UNUSED(val);
-	return -ENOTSUP;
-#endif
-}
-
 /**
  * @brief Retrieve line control for UART.
  *
@@ -1607,25 +1269,6 @@ static inline int z_impl_uart_line_ctrl_set(const struct device *dev,
  */
 __syscall int uart_line_ctrl_get(const struct device *dev, uint32_t ctrl,
 				 uint32_t *val);
-
-static inline int z_impl_uart_line_ctrl_get(const struct device *dev,
-					    uint32_t ctrl, uint32_t *val)
-{
-#ifdef CONFIG_UART_LINE_CTRL
-	const struct uart_driver_api *api =
-		(const struct uart_driver_api *)dev->api;
-
-	if (api->line_ctrl_get == NULL) {
-		return -ENOSYS;
-	}
-	return api->line_ctrl_get(dev, ctrl, val);
-#else
-	ARG_UNUSED(dev);
-	ARG_UNUSED(ctrl);
-	ARG_UNUSED(val);
-	return -ENOTSUP;
-#endif
-}
 
 /**
  * @brief Send extra command to driver.
@@ -1644,25 +1287,6 @@ static inline int z_impl_uart_line_ctrl_get(const struct device *dev,
  */
 __syscall int uart_drv_cmd(const struct device *dev, uint32_t cmd, uint32_t p);
 
-static inline int z_impl_uart_drv_cmd(const struct device *dev, uint32_t cmd,
-				      uint32_t p)
-{
-#ifdef CONFIG_UART_DRV_CMD
-	const struct uart_driver_api *api =
-		(const struct uart_driver_api *)dev->api;
-
-	if (api->drv_cmd == NULL) {
-		return -ENOSYS;
-	}
-	return api->drv_cmd(dev, cmd, p);
-#else
-	ARG_UNUSED(dev);
-	ARG_UNUSED(cmd);
-	ARG_UNUSED(p);
-	return -ENOTSUP;
-#endif
-}
-
 #ifdef __cplusplus
 }
 #endif
@@ -1671,6 +1295,7 @@ static inline int z_impl_uart_drv_cmd(const struct device *dev, uint32_t cmd,
  * @}
  */
 
+#include <zephyr/drivers/uart/internal/uart_impl.h>
 #include <zephyr/syscalls/uart.h>
 
 #endif /* ZEPHYR_INCLUDE_DRIVERS_UART_H_ */
